requirements.txt
watchdog
python-telegram-bot